Fixed Dragonfly.NET compilation error regarding the usage of array
This compilation error was due to some changes made to the include structure in the C++ Standard Library headers since Visual Studio 2015, and the usage of "using namespace std;" in some header file that Dragonfly.NET depends on. The C++/CLI array in the cli namespace should be used, not the one in the std namespace.